✅ Maximum Average Subarray I | Easy | Java | Runtime: 4 ms (43.59%) | Memory: 69.6 MB | 2026-07-13

+++ leetcode/easy/0643-maximum-average-subarray-i/solution.java
@@ -9,7 +9,7 @@ class Solution {
             s=s-nums[i-k]+nums[i];
             m=Math.max(m,s);
         }
-        return m/k;
+        return (double)m/k;
         
     }
 }
